Abstracted away standard javascript list into JsList[T], saving a bunch of repetition

+++ src/main/scala/org/scalajs/dom/extensions/Extensions.scala
@@ -7,7 +7,7 @@ import org.scalajs.dom
  * Used to extend out javascript *Collections to make them usable as normal
  * Scala Seq[*]s
  */
-class EasySeq[T](jsLength: Double, jsApply: Double => T) extends Seq[T] {
+class EasySeq[T](jsLength: Int, jsApply: Int => T) extends Seq[T] {
   def length = jsLength.toInt
 
   def apply(x: Int) = jsApply(x)
